+++ src/app.js
@@ -1,5 +1,6 @@
 import './fixHooksInDev';
 import Uwave from './Uwave';
+import load from './loadingUI';
 import experimentalThemePlugin from './experimentalThemePlugin';
 import youTubeSource from './sources/youtube';
 import soundCloudSource from './sources/soundcloud';
@@ -16,26 +17,7 @@ uw.source(soundCloudSource());
 
 window.uw = uw;
 
-function setLoadingText(txt) {
-  document.querySelector('.LoadingScreen-notice').textContent = text;
-}
-
-setLoadingText('Loading üWave...');
-const longBuildTimer = setTimeout(() => {
-  setLoadingText('Loading is taking a long time, stand by...');
-}, 5000);
-
-uw.build().then(() => {
-  clearTimeout(longBuildTimer);
-  return uw.renderToDOM(document.querySelector('#app'));
-}).then(() => {
-  document.querySelector('#app-loading').innerHTML = '';
-  document.querySelector('#jss').textContent = '';
-}).catch((err) => {
-  setLoadingText(`Error: ${err.message}`);
-  document.querySelector('.LoadingScreen-loader').hidden = true;
-  document.querySelector('.LoadingScreen-warning').hidden = false;
-
+load(uw).catch((err) => {
   setTimeout(() => {
     throw err;
   }, 0);
